+++ CHANGELOG.md
@@ -52,6 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - I18n en DB, permisos docs, UI manager y PDF profesional ([`24e7bc1`](https://github.com/estuardodardon/bifrost/commit/24e7bc1d86ecca250f45b83ca3e2720d8fd93443))
 - Add heartbeat endpoint with severity model and update Postman auth flow ([`4cbe2cf`](https://github.com/estuardodardon/bifrost/commit/4cbe2cf71159fe6855589b9d0c65ff44caa0af5c))
 - Expose endpoint publicly and align payload/contracts ([`4e89d7d`](https://github.com/estuardodardon/bifrost/commit/4e89d7d52831288b07d7039ae9638537a81ab887))
+- :fire: chante default port number ([`dfce6c6`](https://github.com/estuardodardon/bifrost/commit/dfce6c676bce57ab8e4339fc9acfbeebed9e433b))
 
 ### MISCELLANEOUS